Retire legacy attachment production paths
