+++ .env.example
@@ -24,27 +24,8 @@ PUNARO_LOG_LEVEL=info
 # Explicit test/qualification selector. `postgres` requires the compatible
 # schema above and does not import or dual-write the existing SQLite relay.
 # PUNARO_RELAY_STORE=sqlite
-# Directory and permit-issuer values have no safe defaults. If exercising
-# permit issuance, inject all of these from protected service configuration;
-# all IDs/keys except the file path use canonical raw base64url without padding.
-# PUNARO_DIRECTORY_ENABLED=true
-# PUNARO_DIRECTORY_SNAPSHOT_FILE=/secure/punaro/directory.cbor
-# PUNARO_PERMIT_ISSUANCE_ENABLED=true
-# PUNARO_DIRECTORY_AUDIENCE=
-# PUNARO_DIRECTORY_ROOT_KEY_ID=
-# PUNARO_DIRECTORY_ROOT_PUBLIC_KEY=
-# PUNARO_PERMIT_ISSUER_KEY_ID=
-# PUNARO_PERMIT_ISSUER_PRIVATE_KEY_FILE=/secure/punaro/permit-issuer.key
-# PUNARO_PERMIT_MAX_LIFETIME_SECONDS=30
-# PUNARO_PERMIT_MAX_BYTES=1048576
-# PUNARO_PERMIT_MAX_CHUNKS=16
-# PUNARO_PERMIT_MAX_OPERATIONS=4
-# PUNARO_PERMIT_MAX_ACTIVE=32
-# PUNARO_ATTACHMENT_RELAY_ENABLED=false # intentionally rejected until release gates are complete
-# Attachments are disabled by default. Setting this to true intentionally makes
-# the current daemon exit: attachment v2 has not passed its release gates.
-# These future values must be injected from a protected service secret, never
-# committed.
-# PUNARO_ATTACHMENTS_ENABLED=false
-# PUNARO_ATTACHMENT_DEVICE_KEYS_JSON={"device-id":"base64raw-ed25519-public-key"}
-# PUNARO_ATTACHMENT_MEMBERSHIP_JSON=[{"sender":"sender-device","conversation":"conversation-id","recipient":"recipient-device","actions":["create","upload","download","signal"]}]
+# Legacy attachment v2/v3, directory, and permit variables are retired. Their
+# presence, including an empty value or `false`, makes punarod fail startup.
+# Use only the trusted-relay surface and native client documented below.
+# PUNARO_TRUSTED_ATTACHMENTS_ENABLED=false
+# PUNARO_TRUSTED_ATTACHMENT_BLOB_DIR=/absolute/private/punaro/blobs
