+++ internal/config/config.go
@@ -3,8 +3,6 @@ package config
 
 import (
 	"bufio"
-	"crypto/ed25519"
-	"encoding/base64"
 	"fmt"
 	"os"
 	"path/filepath"
@@ -22,24 +20,6 @@ type Config struct {
 	HealthListenAddr            string
 	DataDir                     string
 	LogLevel                    string
-	AttachmentsEnabled          bool
-	AttachmentDeviceKeysJSON    string
-	AttachmentMembershipJSON    string
-	AttachmentV3Enabled         bool
-	AttachmentV3SourceStoreFile string
-	DirectoryEnabled            bool
-	DirectorySnapshotFile       string
-	PermitIssuanceEnabled       bool
-	DirectoryAudience           [32]byte
-	DirectoryRootKeyID          [32]byte
-	DirectoryRootPublicKey      ed25519.PublicKey
-	PermitIssuerKeyID           [32]byte
-	PermitIssuerPrivateKeyFile  string
-	PermitMaxLifetimeSeconds    uint64
-	PermitMaxBytes              uint64
-	PermitMaxChunks             uint64
-	PermitMaxOperations         uint64
-	PermitMaxActive             uint64
 	RelayEnabled                bool
 	RelayMachinesJSON           string
 	RelayStore                  string
@@ -70,6 +50,9 @@ func Load(explicitEnvFile string) (Config, error) {
 			return Config{}, err
 		}
 	}
+	if err := rejectRetiredAttachmentConfiguration(); err != nil {
+		return Config{}, err
+	}
 	level, err := parseLogLevel(value("PUNARO_LOG_LEVEL", "info"))
 	if err != nil {
 		return Config{}, err
@@ -82,40 +65,10 @@ func Load(explicitEnvFile string) (Config, error) {
 		}
 		dataDir = absolute
 	}
-	attachmentsEnabled, err := strconv.ParseBool(value("PUNARO_ATTACHMENTS_ENABLED", "false"))
-	if err != nil {
-		return Config{}, fmt.Errorf("parse PUNARO_ATTACHMENTS_ENABLED: %w", err)
-	}
-	attachmentV3Enabled, err := strconv.ParseBool(value("PUNARO_ATTACHMENT_V3_ENABLED", "false"))
-	if err != nil {
-		return Config{}, fmt.Errorf("parse PUNARO_ATTACHMENT_V3_ENABLED: %w", err)
-	}
 	relayEnabled, err := strconv.ParseBool(value("PUNARO_RELAY_ENABLED", "false"))
 	if err != nil {
 		return Config{}, fmt.Errorf("parse PUNARO_RELAY_ENABLED: %w", err)
 	}
-	deviceKeys := value("PUNARO_ATTACHMENT_DEVICE_KEYS_JSON", "")
-	membership := value("PUNARO_ATTACHMENT_MEMBERSHIP_JSON", "")
-	directoryEnabled, err := strconv.ParseBool(value("PUNARO_DIRECTORY_ENABLED", "false"))
-	if err != nil {
-		return Config{}, fmt.Errorf("parse PUNARO_DIRECTORY_ENABLED: %w", err)
-	}
-	directorySnapshotFile := value("PUNARO_DIRECTORY_SNAPSHOT_FILE", "")
-	permitIssuanceEnabled, err := strconv.ParseBool(value("PUNARO_PERMIT_ISSUANCE_ENABLED", "false"))
-	if err != nil {
-		return Config{}, fmt.Errorf("parse PUNARO_PERMIT_ISSUANCE_ENABLED: %w", err)
-	}
-	directoryAudience := value("PUNARO_DIRECTORY_AUDIENCE", "")
-	directoryRootKeyID := value("PUNARO_DIRECTORY_ROOT_KEY_ID", "")
-	directoryRootPublicKey := value("PUNARO_DIRECTORY_ROOT_PUBLIC_KEY", "")
-	permitIssuerKeyID := value("PUNARO_PERMIT_ISSUER_KEY_ID", "")
-	permitIssuerPrivateKeyFile := value("PUNARO_PERMIT_ISSUER_PRIVATE_KEY_FILE", "")
-	permitMaxLifetimeSeconds := value("PUNARO_PERMIT_MAX_LIFETIME_SECONDS", "")
-	permitMaxBytes := value("PUNARO_PERMIT_MAX_BYTES", "")
-	permitMaxChunks := value("PUNARO_PERMIT_MAX_CHUNKS", "")
-	permitMaxOperations := value("PUNARO_PERMIT_MAX_OPERATIONS", "")
-	permitMaxActive := value("PUNARO_PERMIT_MAX_ACTIVE", "")
-	attachmentV3SourceStoreFile := value("PUNARO_ATTACHMENT_V3_SOURCE_STORE_FILE", "")
 	relayMachines := value("PUNARO_RELAY_MACHINES_JSON", "")
 	relayStore := strings.ToLower(value("PUNARO_RELAY_STORE", "sqlite"))
 	accessIssuer := value("PUNARO_ACCESS_ISSUER", "")
@@ -155,8 +108,7 @@ func Load(explicitEnvFile string) (Config, error) {
 	// The legacy relay origin stays loopback-only. A non-loopback listener is
 	// admitted only when the complete device-ingress policy is validated below
 	// and no legacy route can share that listener.
-	legacyRoutesEnabled := relayEnabled || directoryEnabled || permitIssuanceEnabled || attachmentsEnabled || attachmentV3Enabled
-	if !listener.IsLoopback(listenAddr) && (!deviceAuthEnabled || legacyRoutesEnabled) {
+	if !listener.IsLoopback(listenAddr) && (!deviceAuthEnabled || relayEnabled) {
 		return Config{}, fmt.Errorf("PUNARO_LISTEN_ADDR must be a loopback address until the authenticated public runtime is released")
 	}
 	if deviceAuthEnabled {
@@ -180,21 +132,6 @@ func Load(explicitEnvFile string) (Config, error) {
 	} else if trustedAttachmentBlobDir != "" {
 		return Config{}, fmt.Errorf("PUNARO_TRUSTED_ATTACHMENT_BLOB_DIR requires PUNARO_TRUSTED_ATTACHMENTS_ENABLED")
 	}
-	if attachmentsEnabled && (deviceKeys == "" || membership == "") {
-		return Config{}, fmt.Errorf("attachments require PUNARO_ATTACHMENT_DEVICE_KEYS_JSON and PUNARO_ATTACHMENT_MEMBERSHIP_JSON")
-	}
-	if attachmentV3Enabled && !relayEnabled {
-		return Config{}, fmt.Errorf("v3 attachment runtime requires PUNARO_RELAY_ENABLED")
-	}
-	if attachmentV3Enabled && (attachmentsEnabled || permitIssuanceEnabled) {
-		return Config{}, fmt.Errorf("v3 attachment runtime cannot be enabled with v2 attachment switches")
-	}
-	if attachmentV3Enabled && attachmentV3SourceStoreFile == "" {
-		return Config{}, fmt.Errorf("v3 attachment runtime requires PUNARO_ATTACHMENT_V3_SOURCE_STORE_FILE")
-	}
-	if attachmentV3Enabled && !filepath.IsAbs(attachmentV3SourceStoreFile) {
-		return Config{}, fmt.Errorf("PUNARO_ATTACHMENT_V3_SOURCE_STORE_FILE must be absolute")
-	}
 	if relayEnabled && relayMachines == "" {
 		return Config{}, fmt.Errorf("enabled relay requires PUNARO_RELAY_MACHINES_JSON")
 	}
@@ -204,69 +141,8 @@ func Load(explicitEnvFile string) (Config, error) {
 	if relayStore == "postgres" && (!relayEnabled || !postgresEnabled) {
 		return Config{}, fmt.Errorf("PostgreSQL relay store requires PUNARO_RELAY_ENABLED and PUNARO_POSTGRES_ENABLED")
 	}
-	if relayStore == "postgres" && (directoryEnabled || permitIssuanceEnabled || attachmentV3Enabled || attachmentsEnabled) {
-		return Config{}, fmt.Errorf("PostgreSQL relay store cannot serve superseded attachment or directory routes")
-	}
 	if credentialTransitionEnabled && (!relayEnabled || relayStore != "postgres" || !postgresEnabled || !deviceAuthEnabled) {
 		return Config{}, fmt.Errorf("credential transition requires enabled PostgreSQL relay and device authentication")
-	}
-	if directoryEnabled && !relayEnabled {
-		return Config{}, fmt.Errorf("directory service requires PUNARO_RELAY_ENABLED")
-	}
-	if directoryEnabled && (directorySnapshotFile == "" || !filepath.IsAbs(directorySnapshotFile)) {
-		return Config{}, fmt.Errorf("directory service requires an absolute PUNARO_DIRECTORY_SNAPSHOT_FILE")
-	}
-	var audience, rootKeyID, issuerKeyID [32]byte
-	var rootPublicKey ed25519.PublicKey
-	var maxLifetime, maxBytes, maxChunks, maxOperations, maxActive uint64
-	if permitIssuanceEnabled || attachmentV3Enabled {
-		if !directoryEnabled {
-			return Config{}, fmt.Errorf("permit issuance and v3 attachment runtime require PUNARO_DIRECTORY_ENABLED")
-		}
-		var decodeErr error
-		if audience, decodeErr = decodeFixedBase64URL("PUNARO_DIRECTORY_AUDIENCE", directoryAudience, 32); decodeErr != nil {
-			return Config{}, decodeErr
-		}
-		if rootKeyID, decodeErr = decodeFixedBase64URL("PUNARO_DIRECTORY_ROOT_KEY_ID", directoryRootKeyID, 32); decodeErr != nil {
-			return Config{}, decodeErr
-		}
-		var rootRaw [32]byte
-		if rootRaw, decodeErr = decodeFixedBase64URL("PUNARO_DIRECTORY_ROOT_PUBLIC_KEY", directoryRootPublicKey, ed25519.PublicKeySize); decodeErr != nil {
-			return Config{}, decodeErr
-		}
-		rootPublicKey = append(ed25519.PublicKey(nil), rootRaw[:]...)
-		if issuerKeyID, decodeErr = decodeFixedBase64URL("PUNARO_PERMIT_ISSUER_KEY_ID", permitIssuerKeyID, 32); decodeErr != nil {
-			return Config{}, decodeErr
-		}
-		if permitIssuerPrivateKeyFile == "" || !filepath.IsAbs(permitIssuerPrivateKeyFile) {
-			return Config{}, fmt.Errorf("permit issuance requires an absolute PUNARO_PERMIT_ISSUER_PRIVATE_KEY_FILE")
-		}
-		maxPermitLifetime := uint64(60)
-		if attachmentV3Enabled {
-			maxPermitLifetime = 30
-		}
-		if maxLifetime, decodeErr = parsePositiveUint64("PUNARO_PERMIT_MAX_LIFETIME_SECONDS", permitMaxLifetimeSeconds, maxPermitLifetime); decodeErr != nil {
-			return Config{}, decodeErr
-		}
-		if maxBytes, decodeErr = parsePositiveUint64("PUNARO_PERMIT_MAX_BYTES", permitMaxBytes, 64<<20); decodeErr != nil {
-			return Config{}, decodeErr
-		}
-		if maxChunks, decodeErr = parsePositiveUint64("PUNARO_PERMIT_MAX_CHUNKS", permitMaxChunks, 4096); decodeErr != nil {
-			return Config{}, decodeErr
-		}
-		if maxOperations, decodeErr = parsePositiveUint64("PUNARO_PERMIT_MAX_OPERATIONS", permitMaxOperations, 4096); decodeErr != nil {
-			return Config{}, decodeErr
-		}
-		if maxActive, decodeErr = parsePositiveUint64("PUNARO_PERMIT_MAX_ACTIVE", permitMaxActive, 3*4096+16); decodeErr != nil {
-			return Config{}, decodeErr
-		}
-	}
-	attachmentRelayEnabled, err := strconv.ParseBool(value("PUNARO_ATTACHMENT_RELAY_ENABLED", "false"))
-	if err != nil {
-		return Config{}, fmt.Errorf("parse PUNARO_ATTACHMENT_RELAY_ENABLED: %w", err)
-	}
-	if attachmentRelayEnabled {
-		return Config{}, fmt.Errorf("PUNARO_ATTACHMENT_RELAY_ENABLED is withheld until attachment v2 release gates are complete")
 	}
 	if (accessIssuer == "") != (accessAudience == "") || (accessIssuer == "") != (accessJWKSURL == "" && accessJWKSFile == "") || (accessJWKSURL != "" && accessJWKSFile != "") {
 		return Config{}, fmt.Errorf("PUNARO_ACCESS_ISSUER and PUNARO_ACCESS_AUDIENCE require exactly one of PUNARO_ACCESS_JWKS_URL or PUNARO_ACCESS_JWKS_FILE")
@@ -280,28 +156,34 @@ func Load(explicitEnvFile string) (Config, error) {
 	if !postgresEnabled && postgresDSNFile != "" {
 		return Config{}, fmt.Errorf("PUNARO_POSTGRES_DSN_FILE requires PUNARO_POSTGRES_ENABLED")
 	}
-	return Config{ListenAddr: listenAddr, HealthListenAddr: healthListenAddr, DataDir: dataDir, LogLevel: level, AttachmentsEnabled: attachmentsEnabled, AttachmentDeviceKeysJSON: deviceKeys, AttachmentMembershipJSON: membership, AttachmentV3Enabled: attachmentV3Enabled, AttachmentV3SourceStoreFile: attachmentV3SourceStoreFile, DirectoryEnabled: directoryEnabled, DirectorySnapshotFile: directorySnapshotFile, PermitIssuanceEnabled: permitIssuanceEnabled, DirectoryAudience: audience, DirectoryRootKeyID: rootKeyID, DirectoryRootPublicKey: rootPublicKey, PermitIssuerKeyID: issuerKeyID, PermitIssuerPrivateKeyFile: permitIssuerPrivateKeyFile, PermitMaxLifetimeSeconds: maxLifetime, PermitMaxBytes: maxBytes, PermitMaxChunks: maxChunks, PermitMaxOperations: maxOperations, PermitMaxActive: maxActive, RelayEnabled: relayEnabled, RelayMachinesJSON: relayMachines, RelayStore: relayStore, AccessIssuer: accessIssuer, AccessAudience: accessAudience, AccessJWKSURL: accessJWKSURL, AccessJWKSFile: accessJWKSFile, PostgresEnabled: postgresEnabled, PostgresDSNFile: postgresDSNFile, DeviceAuthEnabled: deviceAuthEnabled, TrustedAttachmentsEnabled: trustedAttachmentsEnabled, TrustedAttachmentBlobDir: trustedAttachmentBlobDir, CredentialTransitionEnabled: credentialTransitionEnabled, IngressMode: ingressMode, PublicURL: publicURL, TrustedLANCIDR: trustedLANCIDR, TrustedLANHTTP: trustedLANHTTP}, nil
+	return Config{ListenAddr: listenAddr, HealthListenAddr: healthListenAddr, DataDir: dataDir, LogLevel: level, RelayEnabled: relayEnabled, RelayMachinesJSON: relayMachines, RelayStore: relayStore, AccessIssuer: accessIssuer, AccessAudience: accessAudience, AccessJWKSURL: accessJWKSURL, AccessJWKSFile: accessJWKSFile, PostgresEnabled: postgresEnabled, PostgresDSNFile: postgresDSNFile, DeviceAuthEnabled: deviceAuthEnabled, TrustedAttachmentsEnabled: trustedAttachmentsEnabled, TrustedAttachmentBlobDir: trustedAttachmentBlobDir, CredentialTransitionEnabled: credentialTransitionEnabled, IngressMode: ingressMode, PublicURL: publicURL, TrustedLANCIDR: trustedLANCIDR, TrustedLANHTTP: trustedLANHTTP}, nil
 }
 
-func decodeFixedBase64URL(name, value string, size int) ([32]byte, error) {
-	var result [32]byte
-	if value == "" || size != len(result) {
-		return result, fmt.Errorf("%s must be canonical base64url %d-byte material", name, size)
+func rejectRetiredAttachmentConfiguration() error {
+	for _, name := range []string{
+		"PUNARO_ATTACHMENTS_ENABLED", "PUNARO_ATTACHMENT_RELAY_ENABLED",
+		"PUNARO_ATTACHMENT_DEVICE_KEYS_JSON", "PUNARO_ATTACHMENT_MEMBERSHIP_JSON",
+		"PUNARO_ATTACHMENT_V3_ENABLED", "PUNARO_ATTACHMENT_V3_SOURCE_STORE_FILE",
+		"PUNARO_ATTACHMENT_RELAY_URL", "PUNARO_ATTACHMENT_DIRECTORY_CHECKPOINT_FILE",
+		"PUNARO_ATTACHMENT_CONTROLLER_JOURNAL", "PUNARO_ATTACHMENT_ARTIFACT_STORE", "PUNARO_ATTACHMENT_OFFER_OUTBOX",
+		"PUNARO_ATTACHMENT_RECIPIENT_ID", "PUNARO_ATTACHMENT_RECIPIENT_GENERATION",
+		"PUNARO_ATTACHMENT_RECIPIENT_SIGNING_PRIVATE_KEY_FILE", "PUNARO_ATTACHMENT_RECIPIENT_HPKE_PRIVATE_KEY_FILE",
+		"PUNARO_ATTACHMENT_SENDER_ID", "PUNARO_ATTACHMENT_SENDER_GENERATION",
+		"PUNARO_ATTACHMENT_SENDER_JOURNAL", "PUNARO_ATTACHMENT_SENDER_SIGNING_PRIVATE_KEY_FILE",
+		"PUNARO_ATTACHMENT_HOST_KEY_SERVICE", "PUNARO_ATTACHMENT_HOST_KEY_ACCOUNT",
+		"PUNARO_ATTACHMENT_HOST_CREDENTIAL_DIRECTORY", "PUNARO_ATTACHMENT_HOST_CREDENTIAL_NAME", "PUNARO_ATTACHMENT_HOST_DPAPI_FILE",
+		"PUNARO_DIRECTORY_ENABLED", "PUNARO_DIRECTORY_SNAPSHOT_FILE",
+		"PUNARO_DIRECTORY_AUDIENCE", "PUNARO_DIRECTORY_ROOT_KEY_ID", "PUNARO_DIRECTORY_ROOT_PUBLIC_KEY",
+		"PUNARO_DIRECTORY_ROOT_PRIVATE_KEY", "PUNARO_DIRECTORY_BINARY", "PUNARO_DIRECTORY_MANIFEST",
+		"PUNARO_PERMIT_ISSUANCE_ENABLED", "PUNARO_PERMIT_ISSUER_KEY_ID", "PUNARO_PERMIT_ISSUER_PRIVATE_KEY_FILE",
+		"PUNARO_PERMIT_MAX_LIFETIME_SECONDS", "PUNARO_PERMIT_MAX_BYTES", "PUNARO_PERMIT_MAX_CHUNKS",
+		"PUNARO_PERMIT_MAX_OPERATIONS", "PUNARO_PERMIT_MAX_ACTIVE",
+	} {
+		if _, present := os.LookupEnv(name); present {
+			return fmt.Errorf("%s is retired; remove legacy attachment v2/v3 configuration and use PUNARO_TRUSTED_ATTACHMENTS_ENABLED", name)
+		}
 	}
-	raw, err := base64.RawURLEncoding.DecodeString(value)
-	if err != nil || len(raw) != size || base64.RawURLEncoding.EncodeToString(raw) != value {
-		return result, fmt.Errorf("%s must be canonical base64url %d-byte material", name, size)
-	}
-	copy(result[:], raw)
-	return result, nil
-}
-
-func parsePositiveUint64(name, value string, maximum uint64) (uint64, error) {
-	parsed, err := strconv.ParseUint(value, 10, 64)
-	if err != nil || parsed == 0 || parsed > maximum {
-		return 0, fmt.Errorf("%s must be an integer from 1 to %d", name, maximum)
-	}
-	return parsed, nil
+	return nil
 }
 
 func value(name, fallback string) string {
