+++ scripts/verify-deployment-files.sh
@@ -4,15 +4,11 @@ set -eu
 unit=deploy/systemd/user/punaro-adapter.service
 example=deploy/systemd/user/punaro-adapter.env.example
 launch_agent=deploy/launchd/punaro-adapter.plist
-snapshot_publisher=scripts/publish-directory-snapshot.sh
-snapshot_publisher_test=scripts/test-publish-directory-snapshot.sh
 adapter_installer=scripts/install-adapter.sh
 client_installer=scripts/install-client.sh
 adapter_installer_test=scripts/test-install-adapter.sh
 server_installer=scripts/install-server.sh
 server_installer_test=scripts/test-install-server.sh
-attachment_provisioner=scripts/provision-attachment-v3.sh
-attachment_provisioner_test=scripts/test-provision-attachment-v3.sh
 attachment_relay_configurer=scripts/configure-attachment-v3-relay.sh
 attachment_relay_configurer_test=scripts/test-configure-attachment-v3-relay.sh
 agent_guidance_installer=scripts/install-agent-guidance.sh
@@ -21,17 +17,16 @@ windows_client_installer=scripts/install-client.ps1
 windows_guidance_installer=scripts/install-agent-guidance.ps1
 windows_client_installer_test=scripts/test-install-client-windows.sh
 windows_adapter_runner=deploy/windows/Run-PunaroAdapter.ps1
-windows_attachment_runner=deploy/windows/Run-PunaroAttachment.ps1
 windows_environment_importer=deploy/windows/Import-PunaroEnvironment.ps1
 
-for path in "$unit" "$example" "$launch_agent" "$snapshot_publisher" "$snapshot_publisher_test" "$adapter_installer" "$client_installer" "$adapter_installer_test" "$server_installer" "$server_installer_test" "$attachment_provisioner" "$attachment_provisioner_test" "$attachment_relay_configurer" "$attachment_relay_configurer_test" "$agent_guidance_installer" "$agent_guidance_installer_test" "$windows_client_installer" "$windows_guidance_installer" "$windows_client_installer_test" "$windows_adapter_runner" "$windows_attachment_runner" "$windows_environment_importer"; do
+for path in "$unit" "$example" "$launch_agent" "$adapter_installer" "$client_installer" "$adapter_installer_test" "$server_installer" "$server_installer_test" "$attachment_relay_configurer" "$attachment_relay_configurer_test" "$agent_guidance_installer" "$agent_guidance_installer_test" "$windows_client_installer" "$windows_guidance_installer" "$windows_client_installer_test" "$windows_adapter_runner" "$windows_environment_importer"; do
 	if [ ! -f "$path" ]; then
 		printf '%s\n' "missing adapter deployment artifact: $path" >&2
 		exit 1
 	fi
 done
 
-for executable in "$adapter_installer" "$client_installer" "$adapter_installer_test" "$server_installer" "$server_installer_test" "$attachment_provisioner" "$attachment_provisioner_test" "$attachment_relay_configurer" "$attachment_relay_configurer_test" "$agent_guidance_installer" "$agent_guidance_installer_test" "$windows_client_installer_test"; do
+for executable in "$adapter_installer" "$client_installer" "$adapter_installer_test" "$server_installer" "$server_installer_test" "$attachment_relay_configurer" "$attachment_relay_configurer_test" "$agent_guidance_installer" "$agent_guidance_installer_test" "$windows_client_installer_test"; do
 	if [ ! -x "$executable" ]; then
 		printf '%s\n' "deployment helper is not executable: $executable" >&2
 		exit 1
@@ -55,49 +50,11 @@ if grep -Eq 'PUNARO_CF_ACCESS_CLIENT_(ID|SECRET)=' "$launch_agent"; then
 	exit 1
 fi
 
-"$snapshot_publisher_test"
 "$adapter_installer_test"
 "$server_installer_test"
-"$attachment_provisioner_test"
 "$attachment_relay_configurer_test"
 "$agent_guidance_installer_test"
 "$windows_client_installer_test"
-
-for expected in \
-	'PUNARO_DIRECTORY_ROOT_PRIVATE_KEY' \
-	'PUNARO_PVE_SSH_TARGET' \
-	'PUNARO_PVE_CONTAINER_ID' \
-	'PUNARO_PVE_SSH_IDENTITY_FILE' \
-	'BatchMode=yes' \
-	'.punaro-directory-publish.lockfile' \
-	'directory_snapshot_publish_already_running' \
-	'PUNARO_PUBLISH_LOCK_HELD' \
-	'fcntl.flock' \
-	'subprocess.run' \
-	'close_fds=True' \
-	'PUNARO_CONTAINER_SNAPSHOT_FILE must be canonical' \
-	'PUNARO_CONTAINER_SNAPSHOT_FILE must not contain parent traversal' \
-	'PUNARO_CONTAINER_SNAPSHOT_FILE must be directly below /etc/punaro/directory' \
-	'/etc/punaro/.punaro-directory-stage' \
-	'install -d -o root -g punaro -m 2750 /etc/punaro/directory' \
-	'chown root:punaro' \
-	'[ ! -L' \
-	'[ ! -L \"\$parent\" ]' \
-	'stat -c %d' \
-	'--ttl 2m' \
-	'PUNARO_CONTAINER_SNAPSHOT_FILE must be below /etc/punaro/directory' \
-	'PUNARO_CONTAINER_SNAPSHOT_FILE contains unsafe characters' \
-	'directory_snapshot_published'; do
-	if ! grep -Fq -- "$expected" "$snapshot_publisher"; then
-		printf '%s\n' "snapshot publisher is missing required safety setting: $expected" >&2
-		exit 1
-	fi
-done
-
-if grep -Fq 'PUNARO_DIRECTORY_ROOT_PRIVATE_KEY=' "$snapshot_publisher"; then
-	printf '%s\n' 'snapshot publisher must not embed a root key value' >&2
-	exit 1
-fi
 
 for expected in \
 	'NoNewPrivileges=yes' \
